+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 All notable changes to this project will be documented in this file. This project adheres to
 [Semantic Versioning](http://semver.org/) and [this changelog format](http://keepachangelog.com/).
 
-## Unreleased
+## [1.0.0] - 2017-09-30
 
 ### Added
 - New collection methods `diff`, `fill`, `fillObject`, `implode` and `intersect`.